Add id to Item object for easier fetching
To make use of a getItem function within the Item Container, an
identification method had to be added to ensure that the right item
could be fetched. Using the existing identification method present in
the item config file naming can just be forwarded to a string within the
Item class.