Fix null added to active item set
Due to the getActiveItems appending null values to the set rather than
cancelling it out, a null value will almost always end up in the set due
to the existence of any disabled items creating a null value. This means
that other null checks that read from the set do not need to be used.