Fix being able to drag item into chest
While the click event prevents most use cases of trying to drag the item
into an inventory other than the player's own, the drag event ensures
that no one trying to bypass the click event check can do so by dragging
the item in instead.

While it completely disables dragging within the player's own inventory
when another inventory is open, it's easier than having to create checks
for every type of inventory as there's no easy way to check which
inventory the item was dragged in as it's based on raw slots. Most
players won't even notice this anyway.