Fix items not being activated on server start
Due to the full filename with extension being forwarded to the
getActiveItem method rather than just the name of the file with serves
as the id, the config is never read from. A substring cut was
implemented to remove the extension from the file to ensure that the
file name is the only thing that is fetched.