Fix duplicate item giving on join
Changed how the check is applied to the inventory to check if the active
item is in the inventory. The looping through the entire inventory and
comparing it to the item is complicated and didn't even work anyway.
Easier to get the inventory and use the method to see if it contains the
ItemStack.