Attach persistent item id to nbt
To more easily identify what item is associated with what persistent
item object, the id is attached to each nbt tag, which makes it easier
when multiple ItemStacks look almost identical at face value but have
different functions to each other. This also makes it easier to get the
command associated with the item.