Rework all existing commands to a subcommand
In reworking all the commands to a more streamlined experience, a
subcommand interface was created which all subcommands can implement.
This is so that the parent command can add the subcommands to a HashMap
and register the subcommands just like the main command is registered
through the main Command Executor.

With this implementation, all subcommands are now children of the main
PersistentItems command, which must be run by calling the main command
and using the name of the subcommand as the first argument.