Fix other metadata affecting prevent item removal
Metadata applied to an item by other players such as lodestone compasses
and potentially bugged stacked items meant that the checks to prevent
items from being removed from the inventory were invalidated. These new
checks simply check if the persistentdatacontainer contains the correct
tag rather than comparing it against the active items list.

This new check also means that if an item is deactivated, it wont be
able to be removed from the inventory by the player regardless. While
this behaviour shouldn't be possible if the item is properly deactivated
and removed from all players' inventories, it should be taken into
consideration if any problems arise.