Refactor project name to Item Binds
This is a little more representative of the name of the project, since
it allows you to bind items to player, and commands to items. The old
name PersistentItems only focuses on the persistence of the item staying
in the inventory between deaths, which is only one aspect of the plugin.